chore: cleaned code
