+++ public/index.html
@@ -29,9 +29,12 @@
   <noscript>You need to enable JavaScript to run this app.</noscript>
 
   <div style="height: 100vh; justify-content: center; align-items: center;">
-    <div id="root" 
-      data-transaction-uuid="" 
-      data-access-token="jF9YS_9mhkJax8xswSOpTI-bu2oxYW472BglAermGe4">
+    <div id="root" data-transaction-uuid="" data-access-token="jF9YS_9mhkJax8xswSOpTI-bu2oxYW472BglAermGe4"
+      data-initial='[
+      {"blockchain": "ada", "name": "Test Asset1", "amount": 1, "image": "ipfs://QmZQTTqDwrXE6RoeqDyiKxpXRD5aJaLoWE4HThMjaFzdQp", "description": "Lorem ipsum lorem ipsum", "attrs": {"hello": "world!"}, "files": [{"name": "test", "src": "ipfs://QmZQTTqDwrXE6RoeqDyiKxpXRD5aJaLoWE4HThMjaFzdQp", "mediaType": null}]},
+      {"blockchain": "ada", "name": "Test Asset2", "amount": 1, "image": "ipfs://QmZQTTqDwrXE6RoeqDyiKxpXRD5aJaLoWE4HThMjaFzdQp", "description": "Lorem ipsum lorem ipsum", "attrs": [{"hello": "world!"}, {"red": "blue"}], "files": [{"name": "test2", "src": "ipfs://QmZQTTqDwrXE6RoeqDyiKxpXRD5aJaLoWE4HThMjaFzdQp", "mediaType": null}]},
+      {"blockchain": "ada", "name": "Test Asset3", "amount": 1, "image": "ipfs://QmZQTTqDwrXE6RoeqDyiKxpXRD5aJaLoWE4HThMjaFzdQp", "description": "Lorem ipsum lorem ipsum", "attrs": {"hello": "world!", "red": "blue", "another": "bites the dust"}, "files": [{"name": "test2", "src": "ipfs://QmZQTTqDwrXE6RoeqDyiKxpXRD5aJaLoWE4HThMjaFzdQp", "mediaType": null}]} 
+    ]'>
     </div>
 
   </div>
