+++ src/components/composites/Asset/AssetForm.tsx
@@ -2,13 +2,9 @@ import React, { useState } from 'react';
 import { Divider, Form as FormDS, Space, Button, Input } from 'antd';
 import { MinusCircleOutlined, PlusOutlined } from '@ant-design/icons';
 import { FileUploader } from 'components/atoms/Input';
-import { insertLineBreaks } from 'utils';
+import { insertLineBreaks, recursiveProperties } from 'utils';
 import { Field, FieldProps } from 'formik';
 import { useFormContext } from 'context/FormContext';
-
-type NestedObject = {
-  [key: string]: NestedObject | string | (NestedObject | string | AssetProps)[];
-};
 
 const AssetForm: React.FC<AssetProps & { index: number }> = ({
   blockchain,
@@ -36,43 +32,12 @@ const AssetForm: React.FC<AssetProps & { index: number }> = ({
 
     const assetUpdate = { ...assetCollection[index] };
 
-    const recursiveProperties: (
-      props: string[],
-      obj?: NestedObject
-    ) => NestedObject | undefined = (props, obj = {}) => {
-      if (!props.length) return obj;
-
-      const propertyKey = props.shift();
-
-      if (propertyKey) {
-        // Check if the propertyKey contains brackets indicating a list
-        const match = propertyKey.match(/^(.+)\[(\d+)\]$/);
-        if (match) {
-          // Extract the list propertyKey and index from the match
-          const listKey = match[1];
-          const idx = parseInt(match[2], 10);
-          // Initialize the list if it doesn't exist
-          if (!Array.isArray(obj[listKey])) {
-            obj[listKey] = [];
-          }
-
-          // Ensure the list element at the specified index is an object
-          // @ts-expect-error the object can be nested object o string
-          obj[listKey][idx] = obj[listKey][idx] || {};
-
-          // RecursivePropertiesly call the function with the updated props and object
-          // @ts-expect-error The object can be a nested object of string
-          recursiveProperties(props, obj[listKey][idx]);
-        } else {
-          obj[propertyKey] = inputElement.value;
-        }
-      }
-
-      return obj;
-    };
-
-    // @ts-expect-error dynamic keys
-    const updatedProperty = recursiveProperties(properties, assetUpdate);
+    const updatedProperty = recursiveProperties(
+      properties,
+      inputElement.value,
+      // @ts-expect-error dynamic keys
+      assetUpdate as NestedObject
+    );
 
     const updated = { ...assetUpdate, ...updatedProperty };
